More work on migrating to long longs for file size.

+++ burn/Source/KWDVDAuthorizer.h
@@ -21,7 +21,7 @@
 	NSDictionary *theme;
 	
 	NSNumber *progressSize;
-	NSInteger fileSize;
+	long long fileSize;
 }
 
 //Standard DVD-Video
